Moved the new functionality from Printer to CodePrinter
